requirements.txt
sklearn
numpy
Pillow
jinja2
